+++ spec/Pim/Bundle/FilterBundle/Filter/ProductValue/ChoiceFilterSpec.php
@@ -15,9 +15,9 @@ use Symfony\Component\Form\FormFactoryInterface;
 
 class ChoiceFilterSpec extends ObjectBehavior
 {
-    function let(FormFactoryInterface $factory, ProductFilterUtility $utility, UserContext $userContext)
+    function let(FormFactoryInterface $factory, ProductFilterUtility $utility, UserContext $userContext, CustomAttributeRepository $repository)
     {
-        $this->beConstructedWith($factory, $utility, $userContext, 'attributeOptionClass');
+        $this->beConstructedWith($factory, $utility, $userContext, 'attributeOptionClass', $repository);
 
         $this->init(
             'foo',
@@ -108,9 +108,10 @@ class ChoiceFilterSpec extends ObjectBehavior
         AttributeRepository $attributeRepository,
         AbstractAttribute $attribute,
         $utility,
-        $factory
+        $factory,
+        $repository
     ) {
-        $utility->getAttribute('data_name_key')->willReturn($attribute);
+        $repository->findOneByCode('data_name_key')->willReturn($attribute);
 
         $factory->create(AjaxChoiceFilterType::NAME, [], [
             'csrf_protection' => false,
@@ -125,5 +126,13 @@ class ChoiceFilterSpec extends ObjectBehavior
         ])->willReturn($form);
 
         $this->getForm()->shouldReturn($form);
+    }
+}
+
+class CustomAttributeRepository extends AttributeRepository
+{
+    public function findOneByCode()
+    {
+        return null;
     }
 }